// frontend/src/app/enums/region.enum.ts
export enum Region {
    CH = 'CH',
    CHFL = 'CHFL',
    AG = 'AG',
    AI = 'AI',
    AR = 'AR',
    BE = 'BE',
    BL = 'BL',
    BS = 'BS',
    FL = 'FL',
    FR = 'FR',
    GE = 'GE',
    GL = 'GL',
    GR = 'GR',
    JU = 'JU',
    LU = 'LU',
    NE = 'NE',
    NW = 'NW',
    OW = 'OW',
    SG = 'SG',
    SH = 'SH',
    SO = 'SO',
    SZ = 'SZ',
    TG = 'TG',
    TI = 'TI',
    UR = 'UR',
    VD = 'VD',
    VS = 'VS',
    ZG = 'ZG',
    ZH = 'ZH'
}